rename modifiers.tsx -> index.tsx
